fix(Application): namespace and file name wrongs
