+++ src/Todo.Application/TodoTasks/Commands/Update/UpdateTaskCommand.cs
@@ -1,7 +1,7 @@
 using Todo.Domain.Enums;
 using Todo.Domain.ValueObjects;
 
-namespace Todo.Application.Commands.Task;
+namespace Todo.Application.TodoTasks.Commands.Update;
 
 public class UpdateTaskCommand
 {
