--- src/Todo.Application/TodoTasks/Queries/GetOne/GetTasksQuery.cs
@@ -1,9 +1,0 @@
-namespace Todo.Application.Queries;
-
-public class GetTasksQuery
-{
-    public Guid UserId { get; init; }
-    public int PageNumber { get; init; }
-    public int PageSize { get; init;}
-
-  }
